feat(tutorial:): add guide for creating interactable slider
The Creating an Interactabel Slider tutorial shows how to use a Linear
Drive to create a simple stepped and free flowing slider along with
the ability to also control some sort of output with the drive position
along the axis.

This implements a simple script to change the opacity of a GameObject
based on the position of the slider.